refactor: View 패키지 구조 재배치
