+++ src/main/java/com/example/demo/view/customer/CustomerView.java
@@ -1,0 +1,31 @@
+package com.example.demo.view.customer;
+
+import com.example.demo.dto.CustomerDto;
+import java.util.List;
+import java.util.UUID;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CustomerView {
+
+    public String readCustomerName() {
+    }
+
+    public int readCustomerAge() {
+    }
+
+    public void printCreateMessage(CustomerDto customerDto) {
+    }
+
+    public void printVoucherList(List<CustomerDto> customerDtoList) {
+    }
+
+    public UUID readCustomerId() {
+    }
+
+    public void printUpdateMessage() {
+    }
+
+    public void printDeleteMessage() {
+    }
+}
