+++ src/main/java/com/example/demo/view/voucher/VoucherView.java
@@ -1,4 +1,4 @@
-package com.example.demo.view;
+package com.example.demo.view.voucher;
 
 import com.example.demo.dto.VoucherDto;
 import com.example.demo.util.CommandType;
